Change Molecule to use size_t
This changes Molecule and its related classes (Atom, Bond,
Ring, Fragment, and Moiety) to use size_t values for indicies
and for returning sizes.

+++ src/chemkit/atom-inline.h
@@ -56,7 +56,7 @@ inline Molecule* Atom::molecule() const
 }
 
 /// Returns the atom's index.
-inline int Atom::index() const
+inline size_t Atom::index() const
 {
     return m_index;
 }
